+++ RELEASES.md
@@ -81,6 +81,7 @@ Upcoming
 - Bugfix monitoring events for input/output-Validators/Wrappers being ignored
 - Add support for multiple versions of the ESB Common Message Header
 - Add SOAP 1.2 support for WsdlXmlValidator and WSDL generator
+- Add interval attribute to scheduler job (run job directly after all adapters are started)